+++ test/e2e/apps/network_partition.go
@@ -211,7 +211,7 @@ var _ = SIGDescribe("Network Partition [Disruptive] [Slow]", func() {
 				defer func() {
 					ginkgo.By(fmt.Sprintf("Unblock traffic from node %s to the master", node.Name))
 					for _, masterAddress := range masterAddresses {
-						framework.UnblockNetwork(host, masterAddress)
+						e2enetwork.UnblockNetwork(host, masterAddress)
 					}
 
 					if ginkgo.CurrentGinkgoTestDescription().Failed {
@@ -226,7 +226,7 @@ var _ = SIGDescribe("Network Partition [Disruptive] [Slow]", func() {
 				}()
 
 				for _, masterAddress := range masterAddresses {
-					framework.BlockNetwork(host, masterAddress)
+					e2enetwork.BlockNetwork(host, masterAddress)
 				}
 
 				ginkgo.By("Expect to observe node and pod status change from Ready to NotReady after network partition")
@@ -599,7 +599,7 @@ var _ = SIGDescribe("Network Partition [Disruptive] [Slow]", func() {
 				defer func() {
 					ginkgo.By(fmt.Sprintf("Unblock traffic from node %s to the master", node.Name))
 					for _, masterAddress := range masterAddresses {
-						framework.UnblockNetwork(host, masterAddress)
+						e2enetwork.UnblockNetwork(host, masterAddress)
 					}
 
 					if ginkgo.CurrentGinkgoTestDescription().Failed {
@@ -611,7 +611,7 @@ var _ = SIGDescribe("Network Partition [Disruptive] [Slow]", func() {
 				}()
 
 				for _, masterAddress := range masterAddresses {
-					framework.BlockNetwork(host, masterAddress)
+					e2enetwork.BlockNetwork(host, masterAddress)
 				}
 
 				ginkgo.By("Expect to observe node and pod status change from Ready to NotReady after network partition")
